rag-streamlit-frontend/requirements.txt
streamlit>=1.48.0
requests>=2.32.0
python-dotenv>=1.1.0
pandas>=2.3.0
plotly>=6.2.0
numpy>=2.0.0 